feat(certificate): moved squared/banner medal images to hook
